Seperated TESTS/API/PWM into three seperate tests.
